requirements.txt
attrs==25.1.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
et_xmlfile==2.0.0
fake-useragent==2.2.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
MouseInfo==0.1.3
numpy==2.3.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pdfkit==1.0.0
psycopg2-binary==2.9.10
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytweening==1.2.0
pytz==2025.2
requests==2.32.4
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
